LLM-HUB-BACKUP/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
stripe==7.3.0
razorpay==1.4.1
openai==1.3.5
mistralai==0.0.7
langchain==0.0.335
transformers==4.35.2
torch==2.1.1
pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.13
bcrypt
google-generativeai
